check all on one display

+++ src/components/frontdash/navs/CheckTabNav.js
@@ -8,7 +8,7 @@ import { doc, setDoc, getDoc, deleteDoc, collection, getCountFromServer } from '
 const CheckTabNav = (props) => {
     const { currentUser, logout, employeeContext, setManagerContext, managerContext } = useAuth();
     const { contextTable } = useTable();
-    const { receiptsList } = props
+    const { receiptsList, allOnOne, setAllOnOne } = props
     const navigate = useNavigate();
     const [ error, setError ] = useState('')
     const [ seatCount, setSeatCount ] = useState('')
@@ -43,13 +43,12 @@ const CheckTabNav = (props) => {
     }
 
     const handleAddSeparate = () => {
-
         const createSeperateReceipt = async () => {
-            if(receiptsList.length >= seatCount){
+            if(receiptsList.length >= seatCount && !allOnOne){
                 props.setFireItAlert('CheckTab more receipts than seats')
             }
 
-            if(receiptsList.length === 0 && receiptsList.length < seatCount){
+            if(receiptsList.length === 0 && receiptsList.length < seatCount && !allOnOne){
                 const receiptRef = 
                         doc(db, 'receipts', `${props.employeeNumber}`, contextTable, 'receipt1')
                 const docSnap = await getDoc(receiptRef)
@@ -64,7 +63,7 @@ const CheckTabNav = (props) => {
                 }
             }
 
-            if(receiptsList.length > 0 && receiptsList.length < seatCount){
+            if(receiptsList.length > 0 && receiptsList.length < seatCount && !allOnOne){
                 const orderedReceipts = receiptsList.slice().sort((a, b) => a - b)
                 let nextReceipt = null
                 
@@ -91,6 +90,10 @@ const CheckTabNav = (props) => {
                     })
                 }
             }
+            
+            if(allOnOne){
+                props.setFireItAlert('CheckTab cancel all on one')
+            }
         }
         createSeperateReceipt()
     }
@@ -101,13 +104,18 @@ const CheckTabNav = (props) => {
             doc(db, 'receipts', `${props.employeeNumber}`, contextTable, `receipt${receiptToRemove}`)
         const removeSeperateReceipt = async () => {
             const docSnap = await getDoc(receiptRef)
-            if(receiptsList.length > 0){
+            if(receiptsList.length > 0 && !allOnOne){
                 if(docSnap.exists()){
                     deleteDoc(receiptRef)
                 }
             }
-            if(receiptsList.length === 0){
+
+            if(receiptsList.length === 0 && !allOnOne){
                 props.setFireItAlert('CheckTab less than zero')
+            }
+
+            if(allOnOne){
+                props.setFireItAlert('CheckTab cancel all on one')
             }
         }
         removeSeperateReceipt()
@@ -115,6 +123,10 @@ const CheckTabNav = (props) => {
 
     const handleChangeTable = () => {
         props.setAlphaNumericPadOpen(true)
+    }
+
+    const handleAllOnOne = () => {
+        setAllOnOne(!allOnOne)
     }
 
     const handleHelp = () => {
@@ -164,7 +176,14 @@ const CheckTabNav = (props) => {
                 <li>
                     <button onClick={handleChangeTable} className='workingButton'>CHNG TBL</button>
                 </li>
-                <li><button className='nonWorkingButton'>ALL ON ONE</button></li>
+                <li>
+                    <button onClick={handleAllOnOne} className='nonWorkingButton'>
+                        {allOnOne
+                            ? 'cancel all on one'
+                            : 'all on one'
+                        }
+                    </button>
+                </li>
                 <li><button className='nonWorkingButton'>% SPLIT</button></li>
                 <li><button onClick={handleTest} className='testButton'>Test</button></li>
                 <li>
